スタート画面UIのバグ直しとStage1の調節
