+++ Assets/Editor/Scripts/GameEnd_M.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEditor;
 
 public class GameEnd_M : MonoBehaviour
 {
