+++ Assets/Masuda/Ingredients/mouse_Lean_red.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:98f4f0173f8a0278b3437b46511c49814559f5fe75178166013b2953c082654e
+size 253572
