+++ Assets/Masuda/Material/Custom_Unlit_Wave.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52fa71b379311b24b83a47460b5f7f96
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
