+++ Assets/Masuda/MouseChange.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f29eb3bdc3bacfb47be47b301b98641a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
