+++ Assets/Masuda/Script_M/GameOff_M.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GameOff_M : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void OnClick()
+    {
+        UnityEditor.EditorApplication.isPlaying = false;
+    }
+}
