+++ Assets/Masuda/Script_M/GameOff_M.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2aca39cba74395458a0d31460668c39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
