--- Assets/Masuda/StoryCS_M/Stage1Clear_M.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b884af3e1e492943b74bdff032fa007
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
