--- Assets/Masuda/StoryCS_M/Stage1Next_M.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Stage1Next_M : MonoBehaviour
-{
-    [SerializeField] GameObject panel;
-    void Start()
-    {
-        panel.SetActive(false);
-    }
-
-    public void OnClick()
-    {
-        panel.SetActive(true);
-    }
-}
