--- Assets/Masuda/StoryCS_M/Stage1Next_M.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dc93b691549fe554290cb3bf1c5fc27b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
