+++ Assets/Masuda/StoryCS_M/Stage1_M.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e13ab1e6f0cb1d74080b76744d310026
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
