+++ Assets/Masuda/StoryCS_M/Stage1_M/Stage1_BuildingBreak_M.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65d75e9e87665f948849d6408090c755
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
