+++ Assets/Masuda/StoryCS_M/Stage1_M/Stage1_Mouse_M.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eaad28d311213304fa75bda11c049f76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
