--- Assets/Masuda/StoryCS_M/Stage1_Mission_M.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 51f68df87b560b2409758398680050ce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
